S^Cit fix2pose into two separate nodes
