+++ eagleye_rt/launch/eagleye_rt.launch.xml
@@ -145,7 +145,9 @@
         </include>
         <include file="$(find-pkg-share eagleye_tf)/launch/tf.launch.xml" if="$(var use_tf)"/>
 
-        <include file="$(find-pkg-share eagleye_fix2pose)/launch/fix2pose.launch.xml">
+        <include file="$(find-pkg-share eagleye_geo_pose_fusion)/launch/geo_pose_fusion.launch.xml">
+        </include>
+        <include file="$(find-pkg-share eagleye_geo_pose_converter)/launch/geo_pose_converter.launch.xml">
         </include>
   </group>
