+++ eagleye_util/geo_pose_converter/package.xml
@@ -1,35 +1,36 @@
 <?xml version="1.0"?>
 <package format="3">
-  <name>eagleye_fix2pose</name>
+  <name>eagleye_geo_pose_converter</name>
   <version>1.0.0</version>
-  <description>fix2pose package</description>
+  <description>geo_pose_converter package</description>
 
   <maintainer email="osamu.sekino@tier4.jp">OsamuSekino</maintainer>
+  <author email="koji.minoda@tier4.jp">Koji Minoda</author>
 
   <license>BSD3</license>
 
-  <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>ament_cmake_auto</buildtool_depend>
   <build_depend>rclcpp</build_depend>
-  <build_depend>rclpy</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>eagleye_msgs</build_depend>
   <build_depend>eagleye_navigation</build_depend>
   <build_depend>tf2_ros</build_depend>
   <build_depend>eagleye_coordinate</build_depend>
+  <build_depend>geographic_msgs</build_depend>
   <build_export_depend>rclcpp</build_export_depend>
-  <build_export_depend>rclpy</build_export_depend>
   <build_export_depend>std_msgs</build_export_depend>
   <build_export_depend>eagleye_msgs</build_export_depend>
   <build_export_depend>eagleye_navigation</build_export_depend>
   <build_export_depend>tf2_ros</build_export_depend>
   <build_export_depend>eagleye_coordinate</build_export_depend>
+  <build_export_depend>geographic_msgs</build_export_depend>
   <exec_depend>rclcpp</exec_depend>
-  <exec_depend>rclpy</exec_depend>
   <exec_depend>std_msgs</exec_depend>
   <exec_depend>eagleye_msgs</exec_depend>
   <exec_depend>eagleye_navigation</exec_depend>
   <exec_depend>tf2_ros</exec_depend>
   <exec_depend>eagleye_coordinate</exec_depend>
+  <exec_depend>geographic_msgs</exec_depend>
   <depend>llh_converter</depend>
   <depend>std_srvs</depend>
