+++ eagleye_util/geo_pose_fusion/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(eagleye_fix2pose)
+project(eagleye_geo_pose_fusion)
 
 # Default to C99
 if(NOT CMAKE_C_STANDARD)
@@ -15,17 +15,13 @@ if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-Wall -Wextra -Wpedantic)
 endif()
 
-if($ENV{ROS_DISTRO} STREQUAL "galactic")
-  add_definitions(-DROS_DISTRO_GALACTIC)
-endif()
-
 find_package(ament_cmake_auto REQUIRED)
 ament_auto_find_build_dependencies()
 
 include_directories(include)
 
-ament_auto_add_executable(fix2pose
-  src/fix2pose.cpp
+ament_auto_add_executable(geo_pose_fusion
+  src/geo_pose_fusion.cpp
 )
 
 install(TARGETS
